% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{getBiomass}
\alias{getBiomass}
\alias{getBiomass.mizerExtensionTemplate}
\title{Get biomass including the plankton component}
\usage{
\method{getBiomass}{mizerExtensionTemplate}(object, ...)
}
\arguments{
\item{object}{A \code{mizerExtensionTemplate} params object.}

\item{...}{Passed to the next method in the dispatch chain.}
}
\value{
A named numeric vector of biomasses (g) with an extra "Plankton"
entry.
}
\description{
Extends \code{\link[mizer:getBiomass]{mizer::getBiomass()}} for \code{mizerExtensionTemplate} objects to
append the plankton component biomass to the standard species biomasses.
}
\details{
Because \code{\link[mizer:plotBiomass]{plotBiomass()}} calls \code{getBiomass()} internally, this single
override is enough to make biomass plots include the plankton without any
further changes.
}
