% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerExtensionTemplate-class.R
\name{mizerExtensionTemplate-class}
\alias{mizerExtensionTemplate-class}
\title{mizerExtensionTemplate extension classes}
\description{
S3 extension classes for \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable dispatch
to the methods defined in this package.
}
\details{
The class names are ordinary entries in the object's S3 class vector. All
extension-specific data lives in
\code{other_params(params)$mizerExtensionTemplate} and in the component
parameters of the "plankton" component (see \code{\link[mizer:setComponent]{setComponent()}}).

Objects of class \code{mizerExtensionTemplate} are created by
\code{\link[=newExtensionTemplateParams]{newExtensionTemplateParams()}}. Objects of class
\code{mizerExtensionTemplateSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerExtensionTemplate} params object.

No class declaration is needed. \code{\link[=newExtensionTemplateParams]{newExtensionTemplateParams()}} records the
extension on the object with \code{\link[mizer:recordExtension]{mizer::recordExtension()}} and then calls
\code{\link[mizer:coerceToExtensionClass]{mizer::coerceToExtensionClass()}}. For example, the params class vector is
\code{c("mizerExtensionTemplate", "MizerParams")}; simulations created by
\code{\link[mizer:project]{project()}} receive
\code{c("mizerExtensionTemplateSim", "MizerSim")} automatically.
}
\keyword{internal}
