% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerExtensionTemplate-package.R
\docType{package}
\name{mizerExtensionTemplate-package}
\alias{mizerExtensionTemplate}
\alias{mizerExtensionTemplate-package}
\title{mizerExtensionTemplate}
\description{
A template for building extension packages for mizer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sizespectrum.org/mizerExtensionTemplate}
  \item \url{https://github.com/sizespectrum/mizerExtensionTemplate}
  \item Report bugs at \url{https://github.com/sizespectrum/mizerExtensionTemplate/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{author@example.com} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID})

Authors:
\itemize{
  \item First Last \email{author@example.com} (\href{https://orcid.org/0000-0003-4092-8228}{ORCID})
}

}
\keyword{internal}
