% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mizerExtensionTemplate-class.R
\name{mizerExtensionTemplate-class}
\alias{mizerExtensionTemplate-class}
\title{mizerExtensionTemplate marker classes}
\description{
S4 marker subclasses of \link[mizer:MizerParams]{MizerParams} and \link[mizer:MizerSim]{MizerSim} that enable S3
dispatch for extension-specific methods defined in this package.
}
\details{
These classes add no new slots. All extension-specific data lives in
\code{other_params(params)$mizerExtensionTemplate} and in the component
parameters of the "plankton" component (see \code{\link[mizer:setComponent]{setComponent()}}).

Objects of class \code{mizerExtensionTemplate} are created by
\code{\link[=newExtensionTemplateParams]{newExtensionTemplateParams()}}. Objects of class
\code{mizerExtensionTemplateSim} are returned automatically by \code{\link[mizer:project]{project()}}
when called on a \code{mizerExtensionTemplate} params object.
}
